pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "secureapi"
version = "0.1.5"
description = "Framework-agnostic API security layer for Django, FastAPI, and Flask"
authors = [
  { name="Vaibhav Patil", email="v7patil77@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"

dependencies = [
  "PyYAML"
]

keywords = ["api security", "authorization", "django", "fastapi", "flask"]

classifiers = [
  "Programming Language :: Python :: 3",
  "Framework :: Django",
  "Framework :: FastAPI",
  "Framework :: Flask",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]